arm: mx6: add initial support for Apalis iMX6
This adds board support for the Toradex module family Apalis iMX6.
The familiy consists of a module with i.MX6 Dual, i.MX6 Quad, i.MX6 Quad IT.

Using a tabulated approach to configure the DDR controller taken from a
patch by Fabio Estevam which achives the same for the Sabre boards.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>