arm: mx6: add initial support for Colibri iMX6
This adds board support for the Toradex module family Colibri iMX6.
The familiy consists of a module with i.MX6 DualLite, i.MX6 Solo, both
with a version for commercial and industrial temperature range.

Using a tabulated approach to configure the DDR controller taken from a
patch by Fabio Estevam which achives the same for the Sabre boards.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>