colibri_imx7/colibri_imx6/apalis_imx6: limit bootm memory
Limit memory used for relocation of FDT or initrd. This is
required to make sure that relocated artifacts are within lowmem.
If fdt_high or initrd_high are not set, U-Boot automatically
relocates artifacts to the end of memory. But this area won't
be part of lowmem and hence will not be accessible by the kernel
during early boot.

With VM split set to 2G/2G (i.MX default), only the 2GB Apalis
iMX6 is affected by that issue. With VM split set to 3G/1G (ARM
default) also modules with 1GB of memory are affected. With the
latter the amount of lowmem will be 760MiB.

The value must also not exceed available memory! Use a safe value
of 512MiB for Apalis and 256MiB for Colibri.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>