arm: dts: imx8mp-toradex-smarc: env: fix boot targets devices
The eMMC is not included in the `boot_targets` environment variable,
meaning the system doesn't attempt to boot from it.

To fix this, update `boot_targets` to include the bootdev mmc2
(eMMC on usdhc3) instead of mmc0 (Wi-Fi sdio on usdhc1).

It will be upstreamed after the bringup is completed.

Upstream-Status: Pending
Fixes: 368fd87e2a30 ("arm: dts: imx8mp-toradex-smarc: sync with DT with Linux")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>