board: toradex: smarc-imx8mp: Kconfig: Fix config block device
With the introduction of commit 38f666759b6a ("arm: dts:
imx8mp-toradex-smarc: add alias of mmc devices"), the eMMC device is
now mapped to mmc0. As a result, the config block device cannot be
not found.

To resolve this issue, set the TDX_CFG_BLOCK_DEV to 0, which is the same
configuration as the SYS_MMC_ENV_DEV.

It will be upstreamed once the bring-up process is completed.

Upstream-Status: Pending
Fixes: 38f666759b6a ("arm: dts: imx8mp-toradex-smarc: add alias of mmc devices")
Signed-off-by: Vitor Soares <vitor.soares@toradex.com>